Remove explicit long cast in NSNumber.LongValue getter
The LongValue property in NSNumber no longer explicitly casts
the result of ObjectiveCRuntime.MsgSendPtr to long. This may
affect the returned type or value depending on MsgSendPtr's
implementation.

+++ Metal.NET/Foundation/NSNumber.cs
@@ -26,7 +26,7 @@ public class NSNumber(nint nativePtr, NativeObjectOwnership ownership) : NativeO
 
     public long LongValue
     {
-        get => (long)ObjectiveCRuntime.MsgSendPtr(NativePtr, NSNumberBindings.LongLongValue);
+        get => ObjectiveCRuntime.MsgSendPtr(NativePtr, NSNumberBindings.LongLongValue);
     }
 
     public ulong UnsignedLongValue
